Create yaml file for easy restart and save logs to a directory
Dump model parameters to a yaml file that will be easily read in later, along with the best model parameters, to have a working model that can be integrated in ARC. With this approach, the user could provide different yaml files and parameter weights in the corresponding ARC file and the model will make TS predictions.

Also save all data for one run, such as log file, yaml file, pymol verifications for early epochs, and plot of training and validation loss, to a directory that organizes runs by date.